Separate the bulk of the create account form to a separate component and add Storybook stories.
Fix a rounding issue where sometimes scrolling to the bottom of the terms and conditions screen wouldn't enable the accept button.
