+++ src/web/components/Core/TermsAndConditions.tsx
@@ -16,7 +16,7 @@ export function TermsAndConditions({ onAccept, onCancel }: TermsAndConditionsPro
   const handleScroll = useCallback((event: UIEvent<HTMLDivElement>) => {
     const { scrollHeight, scrollTop, clientHeight } = event.currentTarget;
     const scrollPos = scrollHeight - scrollTop - clientHeight;
-    if (scrollPos === 0) setScrolledToBottom(true);
+    if (scrollPos < 20) setScrolledToBottom(true);
   }, []);
   return (
     <div className='terms-and-conditions'>
