+++ src/web/components/Input/FormError.tsx
@@ -1,5 +1,16 @@
 import { PropsWithChildren } from 'react';
-import { FieldErrors } from 'react-hook-form';
+import { FieldErrors, FieldValues, UseFormSetError } from 'react-hook-form';
+
+export function setGlobalErrors<TFieldValues extends FieldValues>(
+  setError: UseFormSetError<TFieldValues>,
+  errors: string[]
+) {
+  errors.forEach((m, i) => {
+    setError(`root.serverError-${i}`, {
+      message: m,
+    });
+  });
+}
 
 export function getGlobalErrorsAsArray(root: FieldErrors['root']) {
   if (!root) return [];
